Perfect.
Funciona perfectamente las funciones y con el main acorde. Arreglado los
últimos problemas que tenía y posteriormente comentaré las funciones
para un entendimiento correcto.

+++ printf/ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: danielga <danielga@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/28 12:50:01 by danielga          #+#    #+#             */
-/*   Updated: 2023/01/31 17:46:30 by danielga         ###   ########.fr       */
+/*   Updated: 2023/02/01 20:44:51 by danielga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,18 +26,16 @@ static int	ft_variable(va_list args, char const str)
 	else if (str == 'p')
 	{
 		len += ft_printstr("0x");
-		len += ft_printhex(va_arg(args, unsigned long long), hex);
+		len += ft_printhex(va_arg(args, unsigned long), hex);
 	}
 	else if (str == 'd' || str == 'i')
 		len += ft_printnbr(va_arg(args, int));
 	else if (str == 'u')
 		len += ft_print_unsig_nbr(va_arg(args, unsigned int));
 	else if (str == 'x')
-		len += ft_printhex((unsigned long long)
-				va_arg(args, unsigned long), hex);
+		len += ft_printhex(va_arg(args, unsigned int), hex);
 	else if (str == 'X')
-		len += ft_printhex((unsigned long long)
-				va_arg(args, unsigned long), "0123456789ABCDEF");
+		len += ft_printhex(va_arg(args, unsigned int), "0123456789ABCDEF");
 	else if (str == '%')
 		len += ft_printchar('%');
 	return (len);
