+++ printf/ft_printhex.c
@@ -6,13 +6,13 @@
 /*   By: danielga <danielga@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/20 16:35:14 by danielga          #+#    #+#             */
-/*   Updated: 2023/02/01 17:33:05 by danielga         ###   ########.fr       */
+/*   Updated: 2023/02/01 20:44:00 by danielga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_printhex(unsigned long long p, char *hex)
+int	ft_printhex(unsigned long p, char *hex)
 {
 	int		len;
